starting server refactoring (URL, responses)
