+++ api/webapp/templates/accounts.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Accounts list</h1>
-    <form action="/account" method="get">
+    <form action="/api/account" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
