+++ api/webapp/templates/bookings.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>bookings list</h1>
-    <form action="/booking" method="get">
+    <form action="/api/booking" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
