+++ api/webapp/templates/clients.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Clients list</h1>
-    <form action="/client" method="get">
+    <form action="/api/client" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
