+++ api/webapp/templates/employees.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Employees list</h1>
-    <form action="/employee" method="get">
+    <form action="/api/employee" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
