+++ api/webapp/templates/header.html
@@ -17,15 +17,15 @@
                 <h3 class="float-md-start mb-0">Pet Hotel</h3>
                 <nav class="nav nav-masthead justify-content-center float-md-end">
                     <a class="nav-link active" aria-current="page" href="/">Main</a>
-                    <a class="nav-link" href="/accounts">accounts</a>
-                    <a class="nav-link" href="/users">users</a>
-                    <a class="nav-link" href="/employees">employees</a>
-                    <a class="nav-link" href="/clients">client</a>
-                    <a class="nav-link" href="/pets">pets</a>
-                    <a class="nav-link" href="/hotels">hotels</a>
-                    <a class="nav-link" href="/seats">seats</a>
-                    <a class="nav-link" href="/rooms">rooms</a>
-                    <a class="nav-link" href="/bookings">bookings</a>
+                    <a class="nav-link" href="/api/accounts">accounts</a>
+                    <a class="nav-link" href="/api/users">users</a>
+                    <a class="nav-link" href="/api/employees">employees</a>
+                    <a class="nav-link" href="/api/clients">client</a>
+                    <a class="nav-link" href="/api/pets">pets</a>
+                    <a class="nav-link" href="/api/hotels">hotels</a>
+                    <a class="nav-link" href="/api/seats">seats</a>
+                    <a class="nav-link" href="/api/rooms">rooms</a>
+                    <a class="nav-link" href="/api/bookings">bookings</a>
 
 
                 </nav>
