+++ api/webapp/templates/hotels.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Hotels list</h1>
-    <form action="/hotel" method="get">
+    <form action="/api/hotel" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
