+++ api/webapp/templates/pets.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Pets list</h1>
-    <form action="/pet" method="get">
+    <form action="/api/pet" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
