+++ api/webapp/templates/rooms.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Rooms list</h1>
-    <form action="/room" method="get">
+    <form action="/api/room" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
