+++ api/webapp/templates/seats.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Seats list</h1>
-    <form action="/seat" method="get">
+    <form action="/api/seat" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
