+++ api/webapp/templates/show_account.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/accounts" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to accounts menu</a>
+        <a href="/api/accounts" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to accounts menu</a>
     </p>
 </main>
