+++ api/webapp/templates/show_booking.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/bookings" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to bookings menu</a>
+        <a href="/api/bookings" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to bookings menu</a>
     </p>
 </main>
