+++ api/webapp/templates/show_client.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/clients" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to clients menu</a>
+        <a href="/api/clients" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to clients menu</a>
     </p>
 </main>
