+++ api/webapp/templates/show_employee.html
@@ -7,7 +7,8 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/employees" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to employees menu</a>
+        <a href="/api/employees" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to employees
+            menu</a>
     </p>
 </main>
