+++ api/webapp/templates/show_hotel.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/hotels" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to hotels menu</a>
+        <a href="/api/hotels" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to hotels menu</a>
     </p>
 </main>
