+++ api/webapp/templates/show_pet.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/pets" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to pets menu</a>
+        <a href="/api/pets" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to pets menu</a>
     </p>
 </main>
