+++ api/webapp/templates/show_room.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/rooms" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to rooms menu</a>
+        <a href="/api/rooms" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to rooms menu</a>
     </p>
 </main>
