+++ api/webapp/templates/show_seat.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/seats" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to seats menu</a>
+        <a href="/api/seats" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to seats menu</a>
     </p>
 </main>
