+++ api/webapp/templates/show_user.html
@@ -7,7 +7,7 @@
     <p class="lead">{{ . }}</p>
     <p class="lead">
         <a href="/" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to main page</a>
-        <a href="/users" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to users menu</a>
+        <a href="/api/users" class="btn btn-lg btn-secondary fw-bold border-white bg-white">Back to users menu</a>
     </p>
 </main>
