+++ api/webapp/templates/users.html
@@ -4,7 +4,7 @@
 
 <main class="px-3">
     <h1>Users list</h1>
-    <form action="/user" method="get">
+    <form action="/api/user" method="get">
         <input type="text" name="id" id="id" placeholder="Type ID" class="form-control"><br>
         <button class="btn btn-warning">Search</button>
